+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/RasterOverlay.h
@@ -17,6 +17,7 @@ class CreditSystem;
 class IPrepareRendererResources;
 class RasterOverlayTileProvider;
 class RasterOverlayCollection;
+class RasterOverlayLoadFailureDetails;
 
 /**
  * @brief Options for loading raster overlays.
@@ -63,6 +64,15 @@ struct CESIUM3DTILESSELECTION_API RasterOverlayOptions {
    * the raster overlay maps to approximately 2x2 pixels on the screen.
    */
   double maximumScreenSpaceError = 2.0;
+
+  /**
+   * @brief A callback function that is invoked when a raster overlay resource
+   * fails to load.
+   *
+   * Raster overlay resources include a Cesium ion asset endpoint, any resources
+   * required for raster overlay metadata, or an individual overlay image.
+   */
+  std::function<void(const RasterOverlayLoadFailureDetails&)> loadErrorCallback;
 };
 
 /**
@@ -209,6 +219,12 @@ public:
    * ownership.
    */
   void destroySafely(std::unique_ptr<RasterOverlay>&& pOverlay) noexcept;
+
+protected:
+  void reportError(
+      const CesiumAsync::AsyncSystem& asyncSystem,
+      const std::shared_ptr<spdlog::logger>& pLogger,
+      RasterOverlayLoadFailureDetails&& errorDetails);
 
 private:
   std::string _name;