+++ CesiumAsync/include/CesiumAsync/Impl/unwrapFuture.h
@@ -12,6 +12,10 @@ struct IdentityUnwrapper {
   template <typename Func> static Func unwrap(Func&& f) {
     return std::forward<Func>(f);
   }
+
+  template <typename Func> static Func unwrapShared(Func&& f) {
+    return std::forward<Func>(f);
+  }
 };
 
 template <typename T> struct ParameterizedTaskUnwrapper {
@@ -19,6 +23,11 @@ template <typename T> struct ParameterizedTaskUnwrapper {
     return [f = std::forward<Func>(f)](T&& t) mutable {
       return f(std::move(t))._task;
     };
+  }
+
+  template <typename Func> static auto unwrapShared(Func&& f) {
+    return
+        [f = std::forward<Func>(f)](const T& t) mutable { return f(t)._task; };
   }
 };
 
@@ -38,6 +47,16 @@ template <typename Func, typename T> auto unwrapFuture(Func&& f) {
       ParameterizedTaskUnwrapper<T>>::type::unwrap(std::forward<Func>(f));
 }
 
+template <typename Func, typename T> auto unwrapSharedFuture(Func&& f) {
+  return std::conditional<
+      std::is_same<
+          typename ContinuationReturnType<Func, T>::type,
+          typename RemoveFuture<
+              typename ContinuationFutureType<Func, T>::type>::type>::value,
+      IdentityUnwrapper,
+      ParameterizedTaskUnwrapper<T>>::type::unwrapShared(std::forward<Func>(f));
+}
+
 template <typename Func> auto unwrapFuture(Func&& f) {
   return std::conditional<
       std::is_same<
@@ -46,6 +65,10 @@ template <typename Func> auto unwrapFuture(Func&& f) {
               typename ContinuationFutureType<Func, void>::type>::type>::value,
       IdentityUnwrapper,
       TaskUnwrapper>::type::unwrap(std::forward<Func>(f));
+}
+
+template <typename Func> auto unwrapSharedFuture(Func&& f) {
+  return unwrapFuture(std::forward<Func>(f));
 }
 
 //! @endcond